Move before() implementation to entrypoint
